restructured the codebase to account for the cobra-cli library, the cmd/root.go src that invokes the TUI is now relegated to be summoned by a cobra command. This implies that the default behaviour of the program are the cobra commands.
