+++ internal/ui/root.go
@@ -1,15 +1,14 @@
-package cmd
+package ui
 
 import (
 	"fmt"
 	"os"
 
 	tea "github.com/charmbracelet/bubbletea"
-	"github.com/wingkwong/bootstrap-cli/internal/ui"
 )
 
 func Execute() {
-	p := tea.NewProgram(ui.New())
+	p := tea.NewProgram(New())
 
 	if _, err := p.Run(); err != nil {
 		fmt.Println("Error running bootstrap-cli:", err)
