+++ main.go
@@ -1,9 +1,1 @@
 package main
-
-import (
-	"github.com/wingkwong/bootstrap-cli/cmd"
-)
-
-func main() {
-	cmd.Execute()
-}
